# agent/src/static/openapi_spec.yaml
openapi: 3.0.3
info:
  title: News Processing API
  description: |
    API for processing news text with AI agents.
    
    This API provides endpoints for:
    - Categorizing news text into predefined categories and subcategories
    - Generating multiple-choice quiz questions based on news content
    - Creating concise summaries with key points and entities
    - Processing news text with all methods at once
  version: 1.0.0
  contact:
    name: News Processing API Support

servers:
  - url: http://localhost:8000
    description: Local development server

paths:
  /health:
    get:
      summary: Health check
      description: Returns the health status of the API
      operationId: healthCheck
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  version:
                    type: string
                    example: 1.0.0

  /categorize:
    post:
      summary: Categorize news text
      description: Categorizes news text into predefined categories and subcategories
      operationId: categorizeNews
      requestBody:
        description: News text to categorize
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsTextRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /quiz:
    post:
      summary: Generate quiz question
      description: Generates a multiple-choice quiz question based on news text
      operationId: generateQuiz
      requestBody:
        description: News text to generate a quiz question from
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsTextRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /summarize:
    post:
      summary: Summarize news text
      description: Generates a summary of news text with key points and entities
      operationId: summarizeNews
      requestBody:
        description: News text to summarize
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsTextRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummaryResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /process:
    post:
      summary: Process news text with all methods
      description: Processes news text with all available methods - categorization, quiz generation, and summarization
      operationId: processAll
      requestBody:
        description: News text to process
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsTextRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessAllResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string

components:
  schemas:
    NewsTextRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: The news text to process
          example: SpaceX successfully launched its Starship rocket on a test flight today, marking a significant milestone in the company's efforts to develop a fully reusable spacecraft capable of carrying humans to Mars.
    
    CategoryResponse:
      type: object
      properties:
        main_category:
          type: string
          description: The main category of the news text
          example: Technology
        subcategory:
          type: string
          description: The subcategory within the main category
          example: Space Exploration
        explanation:
          type: string
          description: Brief explanation for the categorization
          example: The news is about SpaceX's rocket launch, which falls under space exploration technology.
    
    QuizChoice:
      type: object
      properties:
        A:
          type: string
          description: First option
          example: The company launched its first manned spacecraft
        B:
          type: string
          description: Second option
          example: Its Starship rocket successfully completed a test flight
        C:
          type: string
          description: Third option
          example: SpaceX introduced a new type of fuel for its rockets
        D:
          type: string
          description: Fourth option
          example: The company inaugurated its new headquarters on Mars
    
    QuizResponse:
      type: object
      properties:
        question:
          type: string
          description: The quiz question
          example: What milestone achievement did SpaceX reach recently?
        choices:
          $ref: '#/components/schemas/QuizChoice'
        correct_answer:
          type: string
          description: The correct answer (A, B, C, or D)
          example: B
        explanation:
          type: string
          description: Brief explanation of why this is the correct answer
          example: The news text mentions that SpaceX successfully launched its Starship rocket on a test flight, marking a significant milestone.
    
    SummaryResponse:
      type: object
      properties:
        summary:
          type: string
          description: Concise summary of the news (1-3 sentences)
          example: SpaceX successfully tested its Starship rocket, reaching an altitude of 250 kilometers before landing back on Earth. The test signifies strides in SpaceX's mission to create a reusable spacecraft for human trips to Mars.
        key_points:
          type: array
          description: Key points from the news text
          items:
            type: string
          example:
            - SpaceX completed a successful test flight of the Starship rocket
            - The rocket reached an altitude of 250 kilometers
            - The test advances SpaceX's efforts to develop a reusable spacecraft for Mars missions
        entities:
          type: array
          description: Important entities mentioned in the news text
          items:
            type: string
          example:
            - SpaceX
            - Starship rocket
            - Mars
    
    ProcessAllResponse:
      type: object
      properties:
        categorization:
          $ref: '#/components/schemas/CategoryResponse'
        quiz:
          $ref: '#/components/schemas/QuizResponse'
        summary:
          $ref: '#/components/schemas/SummaryResponse'
    
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: Failed to parse categorization result
        raw_output:
          type: string
          description: Raw output from the processing agent
          example: Invalid output format